+++ src/main/java/com/raja/lib/invt/repository/LibraryFeeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.raja.lib.invt.model.LibraryFee;
 
-public interface LibraryFeeRepository extends JpaRepository<LibraryFee, Long> {
+public interface LibraryFeeRepository extends JpaRepository<LibraryFee, Integer> {
 }
