+++ src/main/java/com/raja/lib/invt/repository/MemberMonthlyFeesRepository.java
@@ -3,5 +3,5 @@ package com.raja.lib.invt.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.raja.lib.invt.model.MemberMonthlyFees;
 
-public interface MemberMonthlyFeesRepository extends JpaRepository<MemberMonthlyFees, Long> {
+public interface MemberMonthlyFeesRepository extends JpaRepository<MemberMonthlyFees, Integer> {
 }
